File Upload, format checking, and data validation
Updated main.py to include a validate_csv_file() function that performs data validation on a file. This functions is called inside upload_file() .post API endpoint that is responsible for allow the user to upload files from the front-end. The file is uploaded and everything is checked. If it passes the checks, it is saved in the uploads folder. Otherwise it is deleted from the uploads folder